backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
pandas>=2.0,<2.2
reportlab>=4.0
matplotlib>=3.7
Pillow>=10.0
numpy<2
